fix(migrate): rename index to command handler
