--- packages/amplify-migration/API.md
@@ -1,18 +1,0 @@
-## API Report File for "@aws-amplify/migrate"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-// @public (undocumented)
-export type AuthCliInputs = Record<string, unknown>;
-
-// @public (undocumented)
-export function execute(): Promise<void>;
-
-// @public (undocumented)
-export function generateTemplates(fromStack: string, toStack: string): Promise<void>;
-
-// (No @packageDocumentation comment for this package)
-
-```
