+++ packages/amplify-migration/package.json
@@ -2,7 +2,6 @@
   "name": "@aws-amplify/migrate",
   "version": "0.1.0-gen2-migration-test-alpha.0",
   "type": "commonjs",
-  "main": "lib/index.js",
   "bin": "lib/migrate.js",
   "publishConfig": {
     "access": "public"
