+++ packages/amplify-migration/src/commands/gen2/generate-templates/generate-templates_command.ts
@@ -1,5 +1,5 @@
 import { ArgumentsCamelCase, Argv, CommandModule } from 'yargs';
-import { generateTemplates } from '../../../';
+import { generateTemplates } from '../../../command-handlers';
 import assert from 'node:assert';
 
 export interface GenerateTemplatesCommandOptions {
