+++ packages/amplify-migration/src/commands/gen2/start/start_command.ts
@@ -1,5 +1,5 @@
 import { Argv, CommandModule } from 'yargs';
-import { execute } from '../../../';
+import { execute } from '../../../command-handlers';
 
 export type Gen2StartCommandOptions = Record<string, never>;
